Contacting Scryfall API
Pipes can now pull a random card from the scryfall API website, a great MTG API that is really sophisticated. Pipes now can take a command to print an embed on discord for cards. Also now Working on a feature for pipes where users can guess the name of a pulled card.