Pipes now stores spawned cards
Now storing spawned cards so that it doesn't have to be stored locally. Card claiming is now working properly. Checks if the card is in the spawned table. If the name is correct, it will drop the card from the spawned table and then return a confirmation message. Next is to put the card/user relation onto its own table.